requirements.txt
git+https://github.com/yuipanda/telemetry.git@override-timestamp
python-dateutil
